pyproject.toml
[tool.poetry]
name = "avito-parser"
version = "0.1.0"
description = ""
authors = ["MariaZadorina <maria.zadorina@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.3"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
fake-useragent = "^2.1.0"
selenium = "^4.31.0"
fastapi-admin = "^1.0.4"
sqlalchemy = "^2.0.40"
databases = "^0.9.0"
apscheduler = "^3.11.0"
sqladmin = "^0.20.1"
psycopg2-binary = "^2.9.10"
aiohttp = "^3.11.16"
pymysql = "^1.1.1"
python-dotenv = "^1.1.0"
pre-commit = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
